% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw NOAA earthquake data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Data frame containing raw NOAA significant earthquakes data.}
}
\value{
Data frame containing the cleaned data as described in description.
}
\description{
Cleans raw NOAA earthquake data by:
  1/ conforming all variable data types
  2/ uniting the date information columns into a single DATE variable
  3/ cleaning the LOCATION_NAME information to format and keep only
     the specific site location excluding country, state, etc.
}
\examples{
# note this example does not include all the columns in the NOAA dataset,
# but only the ones which eq_clean_data manipulates
raw_df = data.frame(FLAG_TSUNAMI='Tsu', COUNTRY='USA', STATE='',
                    LOCATION_NAME='USA: Pacific Islands',
                    YEAR=2020, MONTH=1, DAY=1, HOUR=NA, MINUTE=NA, SECOND=NA,
                    stringsAsFactors=FALSE)
clean_df = eq_clean_data(raw_df)
}
