% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Name column data}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{Data frame containing column LOCATION_NAME}
}
\value{
Returns data frame with LOCATION_NAME variable cleaned.
}
\description{
Cleans the LOCATION_NAME column by splitting the string at every ':' separator
and keeping the last substring (e.g. 'China: Shenzhen' -> 'Shenzhen').
Finally name is converted to title case.
}
