% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_map}
\alias{eq_map}
\title{Visualize signficant earthquakes}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{Data frame containing cleaned earthquake data.}

\item{annot_col}{Character string of the column name in `df` containing
popup annotation text (e.g. output of eq_create_label or an existing column
of the data frame `df`)}
}
\description{
Visualize a `leaflet` map with locations and data of signficant earthquakes,
based on longitude, latitude, earthquake strength and number of deaths.
}
\examples{
# The following example loads, cleans, and displays earthquakes in Mexico
# occurred after the year 2000:
\dontrun{
require(readr)
require(dplyr)

read_delim('earthquakes.txt', delim = '\\t') \%>\%
  eq_clean_data \%>\%
  filter(COUNTRY == 'MEXICO' & DATE >= '2000-01-01') \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = 'popup_text')
}
}
