% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{geom_timeline}
\alias{geom_timeline}
\alias{GeomTimeline}
\title{Visualize timeline of earthquakes}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)

GeomTimeline
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()`}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data}

\item{position}{Position adjustment}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics, rather than combine}

\item{...}{other arguments passed on to `layer()`}
}
\description{
Visualize the times at which earthquakes occur within certain countries,
along with the magnitude (Richter scale value) and number of deaths associated
with each earthquake.
}
\section{Aesthetics}{

\describe{
  \item{x}{Date}
  \item{group}{(optional) country factor if displaying data for multiple countries;
     multiple countries will be shown on separate timelines}
  \item{size`}{(optional) size of the marker for each earthquake (e.g. magnitude)}
  \item{color}{(optional) color of the marker for each earthquake (e.g. number of deaths)}
  \item{alpha}{(optional) transparency channel for the markers}
}
}

\examples{
# The following example displays earthquake data for the USA from the years 2000 - 2016:

\dontrun{
df \%>\%
  filter('2000-01-01' < DATE & DATE < '2017-01-01',
         COUNTRY == 'USA') \%>\%
  ggplot() +
  geom_timeline(aes(x = DATE, size=EQ_PRIMARY, color=DEATHS)) +
  scale_y_continuous(limits = c(0, 4)) +
  labs(y='', size='Richter scale value', color = '# deaths') +
  theme_timeline()
}
}
\keyword{datasets}
