% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{GeomTimelineLabel}
\title{Visualize timeline of earthquakes, labeling the location of the strongest}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{GeomTimeline}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)

GeomTimelineLabel
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()`}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data}

\item{position}{Position adjustment}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics, rather than combine}

\item{...}{other arguments passed on to `layer()`}
}
\description{
Visualizes the times at which earthquakes occur within certain countries,
along with the magnitude (Richter scale value) and number of deaths associated
with each earthquake. Labels the locations of the strongest N (user specified)
earthquakes.
}
\section{Aesthetics}{

\describe{
  \item{label}{Character string variable with the annotation labels (e.g. LOCATION_NAME)}
  \item{n_max`}{Integer specifying the number of strongest earthquakes to annotate.}
}
}

\examples{
# The following example displays earthquake data for the USA and CHINA
# from the years 2000 - 2016, showing the timeline of each country on a separate,
# and labeling the strongest 5 earthquakes in each country.

\dontrun{
df \%>\%
  filter('2000-01-01' < DATE & DATE < '2017-01-01',
         COUNTRY \%in\% c('USA', 'CHINA')) \%>\%
  ggplot() +
  geom_timeline_label(aes(x=DATE, group=COUNTRY, size=EQ_PRIMARY,
                          color=DEATHS, label=LOCATION_NAME)) +
  scale_y_continuous(limits = c(0, 4)) +
  labs(y='', size='Richter scale value', color = '# deaths') +
  theme_timeline()
}
}
\keyword{datasets}
