% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom for timelines}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\value{
A Geom object that is used within the \code{geom_timeline} function.
}
\description{
\code{GeomTimeline} is a custom Geom for displaying timelines of earthquakes.
}
\note{
This is an internal function that is not directly called by the user.
}
\keyword{datasets}
