% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Geom for timeline labels}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\value{
A Geom object that is used within the \code{geom_timeline_label} function.
}
\description{
\code{GeomTimelineLabel} is a custom Geom for displaying timeline labels.
}
\note{
This is an internal function that is not directly called by the user.
  The aesthetics are as follows:
\itemize{
\item x: A vector of dates.
\item label: A character vector, e.g. the column \code{LOCATION_NAME}
\item n_max (optional): An integer value indicating the maximum number of labels per timeline.
  The \code{n_max} highest values of \code{col_max} will be selected.
\item col_max (optional): The column based on which the \code{n_max} highest values will be selected.
\item y (optional): A factor indicating some stratification in which case
  multiple time lines will be plotted for each level of the factor (e.g. country).
\item colour (optional): A colour used for colouring the vertical lines leading to the label.
}
}
\keyword{datasets}
