% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Custom theme for use with the earthquake timeline plots}
\usage{
theme_timeline()
}
\description{
The theme properly formats the axes, background and gridlines.
}
\examples{
\dontrun{
df \%>\%
  ggplot() +
  geom_timeline_label(aes(x=DATE, group=COUNTRY, size=EQ_PRIMARY,
                          color=DEATHS, label=LOCATION_NAME)) +
  scale_y_continuous(limits = c(0, 4)) +
  labs(y='', size='Richter scale value', color = '# deaths') +
  theme_timeline()
}
}
