% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame or tibble with the following columns:
\code{LOCATION_NAME}, \code{EQ_PRIMARY}, and \code{TOTAL_DEATHS}.}
}
\value{
A character vector in HTML format.
}
\description{
This function creates HTML labels for pop-up windows in leaflet maps.
}
\examples{
# html label pop-up as in figure 2 of module 3
pckg <- getwd()
file <- grep("earthquakes.tsv.gz",
             list.files(path = pckg, all.files = TRUE, full.names = TRUE, recursive = TRUE),
            value = TRUE)

readr::read_delim(file = file, delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
