% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location names}
\usage{
eq_location_clean(loc)
}
\arguments{
\item{loc}{A character string or vector of character strings indicating the location.
The country must be separated from the location name by a colon.}
}
\value{
A character string or vector of character strings wih cleaned location names.
}
\description{
This function strips out the country name (including the colon)
and converts names to title case (as opposed to all caps).
}
\examples{
eq_location_clean(c("MEXICO:  MICHOACAN",
                    "SWITZERLAND:  BASEL",
                    "THAILAND:  BANGKOK"))

}
