% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquakes}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{A data frame or tibble with spatial coordinates (\code{LONGITUDE}, \code{LATITUDE}) and
the earthquake magnitude (\code{EQ_PRIMARY}).}

\item{annot_col}{A column contained in \code{data}
whose information will be displayed in the pop-up window.}
}
\value{
An interactive leaflet map with the earthquake locations.
  The marker size is proportional to the earthquake magnitude (\code{EQ_PRIMARY}).
  When clicking on a marker, a pop-up window will open.
}
\description{
This function creates an interactive leaflet map with the earthquake locations.
}
\examples{
# date pop-up as in figure 1 of module 3
pckg <- getwd()
file <- grep("earthquakes.tsv.gz",
             list.files(path = pckg, all.files = TRUE, full.names = TRUE, recursive = TRUE),
            value = TRUE)

readr::read_delim(file = file, delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")

}
