% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Display timelines}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{A vector of dates.}

\item{y}{(optional) A factor indicating some stratification in which case
multiple time lines will be plotted for each level of the factor (e.g. country).}

\item{colour}{(optional) A numeric vector used for colouring the timeline marker borders.}

\item{fill}{(optional) A numeric vector used for colouring the timeline marker bodies.}

\item{size}{(optional) A numeric vector used for adjusting the size the timeline markers.}

\item{alpha}{(optional) A numeric value indicating the marker's level of transparency.}

\item{shape}{(optional) An integer code (0:25) for one of a set of graphics symbols.
See \code{pch} at \code{?points}.}
}
\value{
A timeline plot with one point symbol for each date.
}
\description{
The \code{geom_timeline} function displays timelines of earthquakes based on the \code{GeomTimeline}.
}
\note{
The function only returns reasonable results for CE (Common Era) dates,
  because the \code{lubridate} package cannot handle BCE (Before Common Era) dates, yet.
}
\examples{
# size and colour as figure 1 of module 2
pckg <- getwd()
file <- grep("earthquakes.tsv.gz",
             list.files(path = pckg, all.files = TRUE, full.names = TRUE, recursive = TRUE),
            value = TRUE)

readr::read_delim(file = file, delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(lubridate::year(DATE) \%in\% 2000:2017 & COUNTRY == "USA") \%>\%
  ggplot(aes(x = DATE, size = EQ_PRIMARY, fill = TOTAL_DEATHS)) +
  geom_timeline() +
  theme_classic() +
  theme(legend.position = "bottom") +
  scale_size_continuous(name = "Richter scale value") +
  scale_fill_continuous(name = "# deaths") +
  guides(size = guide_legend(order = 1),
         fill = guide_colourbar(order = 2))

}
