% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_functions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Geom for timeline labels}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\value{
A Geom object that is used within the \code{geom_timeline_label} function.
}
\description{
\code{GeomTimelineLabel} is a custom Geom for displaying timeline labels.
}
\note{
This is an internal function that is not directly called by the user.
}
\keyword{datasets}
